leadscout/backend/requirements.txt
# LeadScout — Python dependencies
# Install: pip install -r requirements.txt
# Then:    playwright install chromium

# API server
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=12.0        # required by uvicorn for WebSocket support
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<5
python-multipart>=0.0.9
python-dotenv>=1.0.1
supabase>=2.5.0
httpx>=0.28.0

# Browser automation (Google Maps + JustDial)
playwright>=1.42.0

# HTTP requests + HTML parsing (IndiaMart + website checker)
requests>=2.31.0
requests[socks]>=2.31.0
PySocks>=1.7.1
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Data handling
pandas>=2.0.0
